Move queue related methods to separate module.
